Replace Test::{Deep,More,Most} with Test2::V0 in t/
We were already using it in a couple of tests so this just makes it
consistent and reduces the deps needed to install this dist.

Also drop a "use 5.016" from a test, this is the highest use VERSION in
this dist and was causing dzil to set it as the minimum. I'm not sure
what the actual minimum Perl version is for this dist but I don't think
it should be implicitly set from one test file.

xt/ is left alone for now as it doesn't affect end users.